move the util functions to device module
